Fixed Explosion

+++ Assets/Scripts/Building System/Blocks/BuildingBlock.cs
@@ -15,7 +15,7 @@ namespace Building_System.Blocks
         [SerializeField] private NetworkSoundPlayer _soundPlayer;
         [SerializeField] private List<Block> _levels;
 
-        private NetworkVariable<ushort> _hp = new(100, NetworkVariableReadPermission.Everyone,
+        private NetworkVariable<int> _hp = new(100, NetworkVariableReadPermission.Everyone,
             NetworkVariableWritePermission.Owner);
 
         public Block CurrentBlock => _levels[_currentLevel.Value];
@@ -26,8 +26,8 @@ namespace Building_System.Blocks
 
         private DateTime _placingTime;
 
-        public ushort StartHp => _startHp;
-        private ushort _startHp;
+        public int StartHp => _startHp;
+        private int _startHp;
 
         private NetworkVariable<ushort> _currentLevel = new(0, NetworkVariableReadPermission.Everyone,
             NetworkVariableWritePermission.Owner);
@@ -68,7 +68,7 @@ namespace Building_System.Blocks
 
 
         [ServerRpc(RequireOwnership = false)]
-        private void SetHpServerRpc(ushort value)
+        private void SetHpServerRpc(int value)
         {
             _hp.Value = value;
             if (_hp.Value <= 0)
@@ -173,7 +173,7 @@ namespace Building_System.Blocks
         public void GetDamage(int damage, bool playSound = true)
         {
             int hp = _hp.Value - damage;
-            SetHpServerRpc((ushort)hp);
+            SetHpServerRpc(hp);
         }
 
         public int GetHp()
